requirements.txt
streamlit==1.56.0
pandas==3.0.2
plotly==6.7.0
anthropic==0.96.0
python-dotenv==1.2.2
curl_cffi==0.15.0
requests==2.33.1
